+++ src/cordova.ts
@@ -12,7 +12,8 @@ import {ExtensionServer} from './extension/extensionServer';
 import * as Q from "q";
 import {PluginSimulator} from "./extension/simulate";
 import {Telemetry} from './utils/telemetry';
-import {IProjectType, TelemetryHelper} from './utils/telemetryHelper';
+import {TelemetryHelper} from './utils/telemetryHelper';
+import {IProjectType} from './utils/cordovaProjectHelper';
 import {TsdHelper} from './utils/tsdHelper';
 
 let PLUGIN_TYPE_DEFS_FILENAME = "pluginTypings.json";
@@ -94,9 +95,9 @@ export function activate(context: vscode.ExtensionContext): void {
     context.subscriptions.push(vscode.commands.registerCommand('cordova.run',
         () => CordovaCommandHelper.executeCordovaCommand(cordovaProjectRoot, "run")));
     context.subscriptions.push(vscode.commands.registerCommand('cordova.simulate.android',
-        () => launchSimulateCommand({ dir: vscode.workspace.rootPath, target: 'chrome', platform: 'android' })));
+        () => launchSimulateCommand({ dir: vscode.workspace.rootPath, target: 'chrome', platform: 'android'})));
     context.subscriptions.push(vscode.commands.registerCommand('cordova.simulate.ios',
-        () => launchSimulateCommand({ dir: vscode.workspace.rootPath, target: 'chrome', platform: 'ios' })));
+        () => launchSimulateCommand({ dir: vscode.workspace.rootPath, target: 'chrome', platform: 'ios'})));
     context.subscriptions.push(vscode.commands.registerCommand('ionic.prepare',
         () => CordovaCommandHelper.executeCordovaCommand(cordovaProjectRoot, "prepare", true)));
     context.subscriptions.push(vscode.commands.registerCommand('ionic.build',
